[Doc] Correct an example command line and change some context order of the sample_data/README (#918)
* Update README.md to correct a command line to initialize the database

As referred in [Docker's Doc](https://docs.docker.com/engine/reference/run/#volume-shared-filesystems), the `host-src` in `--volume=[host-src:]container-dest[:<options>]` must be an absolute path or a name value.
- If you supply an absolute path for the `host-src`, Docker bind-mounts to the path you specify. 
- If you supply a name, Docker creates a named volume by that name.

**A name value must start with an alphanumeric character, followed by `a-z0-9`, `_` (underscore), `.` (period) or `-` (hyphen). An absolute path starts with a `/` (forward slash).**

* Update README.md to make the doc more reader-friendly

Beautify the layout of the doc.

* Update the start_server.sh to avoid a docker bind-mount bug

Change the `host-src` in bind-mount argument to an absolute path.

* Change `Linux` to `Linux/MacOS`

I agree with @frank-zsy 's PR comment and change the `Linux` to `Linux/MacOS`.